move guts tests to silence a deprecation warning
